requirements.txt
aws-cdk-lib==2.134.0
constructs>=10.0.0,<11.0.0
cdk-nag==2.27.0
boto3==1.28.5
pydantic>=2.3.1
python-dateutil
aws_lambda_powertools
pytest==7.2.0
moto
requests==2.25.1
